# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tour Booking API",
    "version": "1.0.0",
    "description": "SRS-TOUR-2026-v1.0. Values in VND; UTC ISO 8601 timestamps."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1"
    }
  ],
  "paths": {
...
    "/tours/{id}/schedules": {
      "get": {
        "summary": "Các lịch OPEN trong tương lai.",
        "operationId": "get_tours__id__schedules",
        "tags": ["tours"],
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tourId": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "departureAt": {
                                "type": "string",
                                "format": "date-time"
                              },
                              "totalSeats": {
                                "type": "integer",
                                "minimum": 1,
                                "maximum": 10000
                              },
                              "adultPrice": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 99999999
                              },
                              "childPrice": {
                                "type": "integer",
                                "minimum": 0,
                                "maximum": 99999999
                              },
                              "status": {
                                "type": "string",
                                "enum": ["OPEN", "CLOSED"],
                                "default": "OPEN"
                              },
                              "id": {
                                "type": "string",
                                "format": "uuid"
                              },
                              "reservedSeats": {
                                "type": "integer"
                              },
                              "availableSeats": {
                                "type": "integer"
                              },
                              "serverTime": {
                                "type": "string",
                                "format": "date-time"
                              }
                            },
                            "required": [
                              "tourId",
                              "departureAt",
                              "totalSeats",
                              "adultPrice",
                              "childPrice",
                              "id",
                              "reservedSeats",
                              "availableSeats",
                              "serverTime"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        }
                      },
                      "required": ["items", "page", "pageSize", "total"],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["requestId", "timestamp"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "meta"],
                  "additionalProperties": false
                },
                "example": {
                  "data": {
                    "items": [
                      {
                        "id": "22222222-2222-4222-8222-222222222222",
                        "tourId": "11111111-1111-4111-8111-111111111111",
                        "departureAt": "2026-12-15T01:00:00.000Z",
                        "totalSeats": 30,
                        "reservedSeats": 8,
                        "availableSeats": 22,
                        "adultPrice": 3990000,
                        "childPrice": 2490000,
                        "status": "OPEN",
                        "serverTime": "2026-12-01T01:00:00.000Z"
                      }
                    ],
                    "page": 1,
                    "pageSize": 20,
                    "total": 1
                  },
                  "meta": {
                    "requestId": "example-request-id",
                    "timestamp": "2026-12-01T01:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/schedules/{id}/availability": {
      "get": {
        "summary": "Đọc PostgreSQL và thu hồi hold quá hạn trước khi trả chỗ. Không cache.",
        "operationId": "get_schedules__id__availability",
        "tags": ["schedules"],
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "tourId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "departureAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "totalSeats": {
                          "type": "integer",
                          "minimum": 1,
                          "maximum": 10000
                        },
                        "adultPrice": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 99999999
                        },
                        "childPrice": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 99999999
                        },
                        "status": {
                          "type": "string",
                          "enum": ["OPEN", "CLOSED"],
                          "default": "OPEN"
                        },
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "reservedSeats": {
                          "type": "integer"
                        },
                        "availableSeats": {
                          "type": "integer"
                        },
                        "serverTime": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "tourId",
                        "departureAt",
                        "totalSeats",
                        "adultPrice",
                        "childPrice",
                        "id",
                        "reservedSeats",
                        "availableSeats",
                        "serverTime"
                      ],
                      "additionalProperties": false
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": ["requestId", "timestamp"],
                      "additionalProperties": false
                    }
                  },
                  "required": ["data", "meta"],
                  "additionalProperties": false
                },
                "example": {
                  "data": {
                    "id": "22222222-2222-4222-8222-222222222222",
                    "tourId": "11111111-1111-4111-8111-111111111111",
                    "departureAt": "2026-12-15T01:00:00.000Z",
                    "totalSeats": 30,
                    "reservedSeats": 8,
                    "availableSeats": 22,
                    "adultPrice": 3990000,
                    "childPrice": 2490000,
                    "status": "OPEN",
                    "serverTime": "2026-12-01T01:00:00.000Z"
                  },
                  "meta": {
                    "requestId": "example-request-id",
                    "timestamp": "2026-12-01T01:00:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "See error.code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"],
            "additionalProperties": false
          },
          "meta": {
            "type": "object",
            "properties": {
              "requestId": {
                "type": "string"
              },
              "timestamp": {
                "type": "string",
                "format": "date-time"
              }
            },
            "required": ["requestId", "timestamp"],
            "additionalProperties": false
          }
        },
        "required": ["error", "meta"],
        "additionalProperties": false
      }
    }
  }
}